change the navigation based with NavigationComponent
