+++ .idea/misc.xml
@@ -16,6 +16,8 @@
         <entry key="app/src/main/res/layout/detail_user_follows.xml" value="0.3515625" />
         <entry key="app/src/main/res/layout/detail_user_header.xml" value="0.3515625" />
         <entry key="app/src/main/res/layout/detail_user_repo.xml" value="0.3515625" />
+        <entry key="app/src/main/res/layout/fragment_detail_user.xml" value="0.35260416666666666" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.35260416666666666" />
         <entry key="app/src/main/res/layout/item_list_user_horizontal.xml" value="0.35260416666666666" />
         <entry key="app/src/main/res/layout/item_list_user_vertical.xml" value="0.35260416666666666" />
       </map>
