+++ app/build.gradle
@@ -2,6 +2,8 @@ plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
     id 'kotlin-parcelize'
+    id 'androidx.navigation.safeargs'
+    id 'kotlin-android'
 }
 
 android {
@@ -47,6 +49,15 @@ dependencies {
 
     //Circular
     implementation 'de.hdodenhof:circleimageview:3.1.0'
+
+    //Navigation Component
+    implementation "androidx.navigation:navigation-fragment-ktx:2.5.3"
+    implementation "androidx.navigation:navigation-ui-ktx:2.5.3"
+
+    //Restaurant Review With REtrofit
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
+    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
+    implementation "com.squareup.okhttp3:logging-interceptor:4.10.0"
 
     //coil image loader
     implementation 'io.coil-kt:coil:2.2.1'
