+++ app/src/main/java/com/thatnawfal/githubuser/presentation/logic/UserViewModel.kt
@@ -1,0 +1,5 @@
+package com.thatnawfal.githubuser.presentation.logic
+
+class UserViewModel {
+
+}
