+++ app/src/main/java/com/thatnawfal/githubuser/presentation/ui/home/MainActivity.kt
@@ -2,9 +2,14 @@ package com.thatnawfal.githubuser.presentation.ui.home
 
 import android.annotation.SuppressLint
 import android.content.Intent
+import android.content.pm.ActivityInfo
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.thatnawfal.githubuser.R
 import com.thatnawfal.githubuser.data.model.UserModel
 import com.thatnawfal.githubuser.databinding.ActivityMainBinding
@@ -13,70 +18,25 @@ import com.thatnawfal.githubuser.presentation.ui.profile.DetailUserActivity
 
 class MainActivity : AppCompatActivity() {
 
-    companion object {
-        const val EXTRA_KEY = "extra_key"
-    }
-
     private lateinit var binding : ActivityMainBinding
-    private val adapter : UserAdapter by lazy {
-        UserAdapter()
-    }
+    private lateinit var navController: NavController
 
+    @SuppressLint("SourceLockedOrientationActivity")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        initRecyclerView()
+        requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.navContainer)
+                as NavHostFragment
+
+        navController = navHostFragment.navController
     }
 
-    @SuppressLint("Recycle")
-    private fun addingDataToList() {
-        val dataAvatar = resources.obtainTypedArray(R.array.avatar)
-        val dataName = resources.getStringArray(R.array.name)
-        val dataUsername = resources.getStringArray(R.array.username)
-        val dataRepositoies = resources.getIntArray(R.array.repository)
-        val dataFollowings = resources.getIntArray(R.array.following)
-        val dataFollowers = resources.getIntArray(R.array.followers)
-        val dataLocation = resources.getStringArray(R.array.location)
-        val dataCompany = resources.getStringArray(R.array.company)
 
-        for (i in dataName.indices) {
-            val dataUser = UserModel(
-                dataAvatar.getResourceId(i, -1),
-                dataCompany[i],
-                dataFollowers[i],
-                dataFollowings[i],
-                dataLocation[i],
-                dataName[i],
-                dataRepositoies[i],
-                dataUsername[i]
-            )
-            adapter.addItem(dataUser)
-        }
 
-    }
-
-    private fun initRecyclerView() {
-        binding.rvUserList.setHasFixedSize(true)
-        binding.rvUserList.layoutManager = LinearLayoutManager(this,
-            LinearLayoutManager.VERTICAL, false)
-        binding.rvUserList.adapter = adapter
-
-        addingDataToList()
-
-        adapter.itemClicked(object : UserAdapter.OnItemClickedCallback{
-            override fun itemClicked(item: UserModel) {
-                showSelectedItem(item)
-            }
-        })
-    }
-
-    private fun showSelectedItem(item: UserModel) {
-        val intent = Intent(this, DetailUserActivity::class.java)
-        intent.putExtra(EXTRA_KEY, item)
-        startActivity(intent)
-    }
 
 
 }
