+++ app/src/main/java/com/thatnawfal/githubuser/utils/Helper.kt
@@ -1,0 +1,5 @@
+package com.thatnawfal.githubuser.utils
+
+object Helper {
+    const val api_endpoint = "https://api.github.com/users"
+}
