+++ app/src/main/res/layout/activity_detail_user.xml
@@ -70,5 +70,4 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-
 </ScrollView>
