+++ app/src/main/res/layout/fragment_detail_user.xml
@@ -1,0 +1,78 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".presentation.ui.profile.DetailUserFragment">
+
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+            <include
+                android:id="@+id/follows_detail_user"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toBottomOf="@id/view_bottom_align"
+                layout="@layout/detail_user_follows"/>
+
+            <View
+                android:id="@+id/view_bottom_align"
+                android:layout_width="1dp"
+                android:layout_height="1dp"
+                android:layout_margin="32dp"
+                app:layout_constraintBottom_toBottomOf="@id/header_detail_user"
+                app:layout_constraintStart_toStartOf="@id/header_detail_user"/>
+
+            <include
+                android:id="@+id/header_detail_user"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                layout="@layout/detail_user_header" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/submission_dicoding_fundamental"
+                android:layout_margin="64dp"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/follows_detail_user"/>
+
+            <com.google.android.material.floatingactionbutton.FloatingActionButton
+                android:id="@+id/detailFb_toGithub"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@drawable/image_github"
+                android:layout_marginTop="256dp"
+                android:layout_marginEnd="16dp"
+                android:layout_marginBottom="16dp"
+                app:tint="@color/white"
+                app:backgroundTint="@color/black_80"
+                android:contentDescription="@string/share_detail_layout"
+                app:layout_constraintTop_toTopOf="@id/follows_detail_user"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"/>
+
+            <ImageButton
+                android:id="@+id/back_button_detail"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@drawable/ic_back"
+                android:background="@color/transparant"
+                android:layout_margin="16dp"
+                android:elevation="9dp"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                android:contentDescription="@string/back_button" />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </ScrollView>
+
+</FrameLayout>
