+++ app/src/main/res/values/strings.xml
@@ -128,4 +128,6 @@
         <item>@drawable/user9</item>
         <item>@drawable/user10</item>
     </integer-array>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
